+++ src/main/java/com/warhammer/model/UnitSummary.java
@@ -1,0 +1,25 @@
+package com.warhammer.model;
+
+public class UnitSummary {
+
+    private final int id;
+    private final String name;
+
+    public UnitSummary(int id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public UnitSummary (Unit unit){
+        this.id = unit.getId();
+        this.name = unit.getName();
+    }
+}