+++ src/main/java/combat/AllocationStrategy.java
@@ -1,6 +1,0 @@
-package combat;
-
-public enum AllocationStrategy {
-    WORST_SAVE_FIRST,
-    BEST_SAVE_FIRST
-}