+++ src/main/java/view/PrintLevel.java
@@ -1,7 +1,0 @@
-package view;
-
-public enum PrintLevel {
-    SUMMARY,
-    DETAILED,
-    WEAPON_SUMMARY
-}