+++ changelog.md
@@ -27,7 +27,7 @@ Version 0.4:
 Version 0.3:
 + Refactored Result Classes again with an abstract class
 + BattleResult now has a list of phases
-+ Main class that has hardcoded attacker and defender
++ com.warhammer.Main class that has hardcoded attacker and defender
 + Printer class for console output
 
 Version 0.2: